requirements.txt
requests>=2.32.0
pandas>=2.2.2
pyarrow>=16.1.0
duckdb>=1.0.0
pyyaml>=6.0.1
typer>=0.12.3
tenacity>=8.4.2
lxml>=5.2.2
beautifulsoup4>=4.12.3
streamlit>=1.36.0
